@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.3
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -30,51 +30,51 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
 
   # FIREBASE PACKAGES
-  firebase_core: 1.1.0
-  firebase_messaging: 9.1.3
+  firebase_core: ^1.12.0
+  firebase_messaging: ^11.2.6
 
   # PUSH NOTIFICATION
-  onesignal_flutter: 2.6.3
+  onesignal_flutter: ^3.2.8
 
   # STORAGE
-  shared_preferences: ^2.0.5
+  shared_preferences: ^2.0.13
 
   # UI
-  cupertino_icons: ^1.0.3
+  cupertino_icons: ^1.0.4
   flutter_spinkit: ^5.1.0
-  page_transition: ^2.0.4
+  page_transition: ^2.0.5
   #modal_progress_hud: ^0.1.3
   flutter_inappwebview: ^5.3.2
 
   # ADS
-  google_mobile_ads: 0.12.1+1
+  google_mobile_ads: ^1.0.1
 
   # CONNECTIVITY
   http: ^0.13.4
   share: ^2.0.4
   launch_review: ^3.0.1
-  url_launcher: ^6.0.12
-  store_redirect: ^2.0.0
+  url_launcher: ^6.0.18
+  store_redirect: ^2.0.1
   connectivity: ^3.0.6
   uni_links: ^0.5.1
 
   # TOOLS
-  google_fonts: ^2.1.0
+  google_fonts: ^2.3.1
   package_info: ^2.0.2
   android_intent: ^2.0.2
   global_configuration: ^2.0.0-nullsafety.1
   #flutter_downloader: 1.6.1
   flutter_barcode_scanner: ^2.0.0
-  flutter_speed_dial: ^4.6.6
+  flutter_speed_dial: ^5.0.0+1
 
   # STATE MANAGEMENT
-  provider: ^6.0.1
-  path_provider: ^2.0.6
+  provider: ^6.0.2
+  path_provider: ^2.0.8
 
   # MAP
-  location: 2.5.2
-  geolocator: ^5.3.2+2
-  webview_flutter: ^2.1.2
+  location: ^4.3.0
+  geolocator: ^8.1.1
+  webview_flutter: ^3.0.0
 
 dev_dependencies:
   flutter_test:
