@@ -1,16 +1,15 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class RaisedGradientButton extends StatelessWidget {
   final Widget child;
-  final Gradient gradient;
+  final Gradient? gradient;
   final double width;
   final double height;
-  final Function onPressed;
+  final Function? onPressed;
 
   const RaisedGradientButton({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
     this.gradient,
     this.width = double.infinity,
     this.height = 50.0,
@@ -27,7 +26,7 @@ class RaisedGradientButton extends StatelessWidget {
           borderRadius: BorderRadius.all(Radius.circular(80.0)),
           boxShadow: [
             BoxShadow(
-              color: Colors.grey[500],
+              color: Colors.grey[500]!,
               offset: Offset(0.0, 1.5),
               blurRadius: 1.5,
             ),
@@ -35,7 +34,7 @@ class RaisedGradientButton extends StatelessWidget {
       child: Material(
         color: Colors.transparent,
         child: InkWell(
-            onTap: onPressed,
+            onTap: onPressed as void Function()?,
             child: Center(
               child: child,
             )),
