@@ -1,20 +1,16 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-import 'dart:isolate';
 import 'dart:math' as math;
-import 'dart:typed_data';
-import 'dart:ui';
-import 'package:flutter_speed_dial/flutter_speed_dial.dart';
-import 'package:flyweb/src/models/floating.dart';
-import 'package:http/http.dart' as http;
 
 import 'package:android_intent/android_intent.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart' hide Page;
-import 'package:flutter/services.dart';
 import 'package:flutter_barcode_scanner/flutter_barcode_scanner.dart';
+import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:flutter_speed_dial/flutter_speed_dial.dart';
+import 'package:flyweb/i18n/AppLanguage.dart';
 import 'package:flyweb/i18n/i18n.dart';
 import 'package:flyweb/src/elements/DrawerListTitle.dart';
 import 'package:flyweb/src/elements/Loader.dart';
@@ -23,35 +19,36 @@ import 'package:flyweb/src/enum/connectivity_status.dart';
 import 'package:flyweb/src/helpers/HexColor.dart';
 import 'package:flyweb/src/helpers/OneSignalHelper.dart';
 import 'package:flyweb/src/helpers/SharedPref.dart';
+import 'package:flyweb/src/models/floating.dart';
 import 'package:flyweb/src/models/menu.dart';
 import 'package:flyweb/src/models/navigationIcon.dart';
 import 'package:flyweb/src/models/page.dart';
 import 'package:flyweb/src/models/settings.dart';
 import 'package:flyweb/src/pages/LanguageScreen.dart';
+import 'package:flyweb/src/pages/MosqueSearchScreen.dart';
 import 'package:flyweb/src/pages/PageScreen.dart';
 import 'package:flyweb/src/pages/WebScreen.dart';
 import 'package:flyweb/src/position/PositionOptions.dart';
 import 'package:flyweb/src/position/PositionResponse.dart';
+import 'package:flyweb/src/services/mosque_manager.dart';
+import 'package:flyweb/src/services/settings_manager.dart';
+import 'package:flyweb/src/services/theme_manager.dart';
 import 'package:flyweb/src/themes/UIImages.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
+import 'package:http/http.dart' as http;
 import 'package:launch_review/launch_review.dart';
 import 'package:location/location.dart' hide LocationAccuracy;
 import 'package:package_info/package_info.dart';
 import 'package:page_transition/page_transition.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:share/share.dart';
 import 'package:store_redirect/store_redirect.dart';
 import 'package:uni_links/uni_links.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
-
 import 'AboutScreen.dart';
-
-import 'package:flyweb/src/services/theme_manager.dart';
 
 GlobalKey<_WebViewScreen> key0 = GlobalKey();
 GlobalKey<_WebViewScreen> key1 = GlobalKey();
@@ -76,10 +73,9 @@ List<StreamController<int>> listStream = [
 ];
 
 class HomeScreen extends StatefulWidget {
-  final String url;
   final Settings settings;
 
-  const HomeScreen(this.url, this.settings);
+  const HomeScreen(this.settings);
 
   @override
   State<StatefulWidget> createState() {
@@ -88,21 +84,20 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreen extends State<HomeScreen>
-    with TickerProviderStateMixin, SingleTickerProviderStateMixin {
+    with SingleTickerProviderStateMixin {
   SharedPref sharedPref = SharedPref();
-  Settings settings = new Settings();
 
-  BannerAd _bannerAd;
+  BannerAd? _bannerAd;
   bool _isBannerAdReady = false;
-  InterstitialAd _interstitialAd;
+  InterstitialAd? _interstitialAd;
   bool _isInterstitialAdReady = false;
 
   static GlobalKey<ScaffoldState> _scaffoldKey = new GlobalKey<ScaffoldState>();
   String url = "";
 
-  Uri _initialUri;
-  Uri _latestUri;
-  StreamSubscription _sub;
+  Uri? _initialUri;
+  Uri? _latestUri;
+  StreamSubscription? _sub;
   bool goToWeb = true;
 
   List<StreamSubscription<Position>> webViewGPSPositionStreams = [];
@@ -114,41 +109,42 @@ class _HomeScreen extends State<HomeScreen>
     Factory<PanGestureRecognizer>(() => PanGestureRecognizer()),
   ].toSet();
 
-  PackageInfo _packageInfo = PackageInfo(
-    appName: 'Unknown',
-    packageName: 'Unknown',
-    version: 'Unknown',
-    buildNumber: 'Unknown',
-  );
+  // PackageInfo _packageInfo = PackageInfo(
+  //   appName: 'Unknown',
+  //   packageName: 'Unknown',
+  //   version: 'Unknown',
+  //   buildNumber: 'Unknown',
+  // );
 
-  TabController tabController;
+  TabController? tabController;
   int _currentIndex = 0;
 
   @override
   void initState() {
     super.initState();
-    _initPackageInfo();
+    // _initPackageInfo();
 
     _handleIncomingLinks();
 
     tabController = new TabController(
-        initialIndex: 0,
-        length: widget.settings.tabNavigationEnable == "1"
-            ? widget.settings.tab.length
-            : 1,
-        vsync: this);
-    tabController.addListener(_handleTabSelection);
+      initialIndex: 0,
+      length: widget.settings.tabNavigationEnable == "1"
+          ? widget.settings.tab!.length
+          : 1,
+      vsync: this,
+    );
+    tabController!.addListener(_handleTabSelection);
 
     if (widget.settings.adBanner == "1") {
       String adBannerId = Platform.isAndroid
-          ? widget.settings.admobKeyAdBanner
-          : widget.settings.admobKeyAdBannerIos;
+          ? widget.settings.admobKeyAdBanner!
+          : widget.settings.admobKeyAdBannerIos!;
       // TODO: Initialize _bannerAd
       _bannerAd = BannerAd(
         adUnitId: adBannerId,
         request: AdRequest(),
         size: AdSize.banner,
-        listener: AdListener(
+        listener: BannerAdListener(
           onAdLoaded: (_) {
             setState(() {
               _isBannerAdReady = true;
@@ -162,40 +158,42 @@ class _HomeScreen extends State<HomeScreen>
         ),
       );
 
-      _bannerAd.load();
+      _bannerAd!.load();
     }
 
     if (widget.settings.adInterstitial == "1") {
-      String adInterstitialId = Platform.isAndroid
+      String? adInterstitialId = Platform.isAndroid
           ? widget.settings.admobKeyAdInterstitial
           : widget.settings.admobKeyAdInterstitialIos;
-      _interstitialAd = InterstitialAd(
-        adUnitId: adInterstitialId,
-        request: AdRequest(),
-        listener: AdListener(
-          onAdLoaded: (_) {
-            _isInterstitialAdReady = true;
-          },
-          onAdFailedToLoad: (ad, err) {
-            print('Failed to load an interstitial ad: ${err.message}');
-            _isInterstitialAdReady = false;
-            ad.dispose();
-          },
-        ),
-      );
 
       Timer.periodic(
-          new Duration(seconds: int.parse(widget.settings.admobDealy)),
+          new Duration(seconds: int.parse(widget.settings.admobDealy!)),
           (timer) {
-        _interstitialAd?.load();
-        _interstitialAd?.show();
+        InterstitialAd.load(
+          adUnitId: adInterstitialId!,
+          request: AdRequest(),
+          adLoadCallback: InterstitialAdLoadCallback(
+            onAdLoaded: (ad) {
+              _isInterstitialAdReady = true;
+              _interstitialAd = ad;
+              ad.show();
+            },
+            onAdFailedToLoad: (err) {
+              print('Failed to load an interstitial ad: ${err.message}');
+              _isInterstitialAdReady = false;
+              // ad.dispose();
+            },
+          ),
+        );
+        // _interstitialAd?.load();
+        // _interstitialAd?.show();
       });
     }
   }
 
   void _handleIncomingLinks() {
     if (!kIsWeb) {
-      _sub = uriLinkStream.listen((Uri uri) async {
+      _sub = uriLinkStream.listen((Uri? uri) async {
         if (!mounted) return;
         print('got uri: $uri');
         setState(() {
@@ -212,16 +210,20 @@ class _HomeScreen extends State<HomeScreen>
             final result = await Navigator.push(
                 context,
                 PageTransition(
-                    type: PageTransitionType.rightToLeft,
-                    child: WebScreen(link, widget.settings)));
+                  type: pageTransitionAnimation(context),
+                  child: WebScreen(link),
+                ));
 
             setState(() {
               goToWeb = true;
             });
           }
         } else {
-          key0.currentState._webViewController
-              ?.loadUrl(urlRequest: URLRequest(url: Uri.parse(link)));
+          key0.currentState!._webViewController?.loadUrl(
+            urlRequest: URLRequest(
+              url: Uri.parse(link),
+            ),
+          );
         }
       }, onError: (Object err) {
         if (!mounted) return;
@@ -234,9 +236,7 @@ class _HomeScreen extends State<HomeScreen>
   }
 
   _handleTabSelection() {
-    setState(() {
-      _currentIndex = tabController.index;
-    });
+    setState(() => _currentIndex = tabController!.index);
   }
 
   @override
@@ -251,12 +251,12 @@ class _HomeScreen extends State<HomeScreen>
     super.dispose();
   }
 
-  Future<void> _initPackageInfo() async {
-    final PackageInfo info = await PackageInfo.fromPlatform();
-    setState(() {
-      _packageInfo = info;
-    });
-  }
+  // Future<void> _initPackageInfo() async {
+  //   final PackageInfo info = await PackageInfo.fromPlatform();
+  //   setState(() {
+  //     _packageInfo = info;
+  //   });
+  // }
 
   Future<InitializationStatus> _initGoogleMobileAds() {
     // TODO: Initialize Google Mobile Ads SDK
@@ -265,76 +265,87 @@ class _HomeScreen extends State<HomeScreen>
 
   @override
   Widget build(BuildContext context) {
-    MediaQueryData mediaQueryData = MediaQuery.of(context);
-    var bottomPadding = mediaQueryData.padding.bottom;
+    final appLanguage = context.watch<AppLanguage>();
+    final mosqueManager = context.watch<MosqueManager>();
+    final settingsManager = context.read<SettingsManager>();
+    final settings = settingsManager.settings;
+
+    var url =
+        'https://mawaqit.net/${appLanguage.appLocal.languageCode}/id/${mosqueManager.mosqueId}?view=desktop';
+
+    print(url);
+    var bottomPadding = MediaQuery.of(context).padding.bottom;
     var connectionStatus = Provider.of<ConnectivityStatus>(context);
 
     var themeProvider = Provider.of<ThemeNotifier>(context);
 
     if (connectionStatus == ConnectivityStatus.Offline)
-      return _offline(bottomPadding);
+      return _offline(bottomPadding, settings);
 
     final _oneSignalHelper = OneSignalHelper();
     Future<void> _listenerOneSignal() async {
       print(_oneSignalHelper.url);
-      if (widget.settings.tabNavigationEnable == "1") {
+      if (settings.tabNavigationEnable == "1") {
         if (goToWeb) {
           setState(() {
             goToWeb = false;
           });
           final result = await Navigator.push(
-              context,
-              PageTransition(
-                  type: PageTransitionType.rightToLeft,
-                  child: WebScreen(_oneSignalHelper.url, widget.settings)));
+            context,
+            PageTransition(
+              type: pageTransitionAnimation(context),
+              child: WebScreen(_oneSignalHelper.url),
+            ),
+          );
 
           setState(() {
             goToWeb = true;
           });
         }
       } else {
-        key0.currentState._webViewController?.loadUrl(
-            urlRequest: URLRequest(url: Uri.parse(_oneSignalHelper.url)));
+        key0.currentState!._webViewController?.loadUrl(
+          urlRequest: URLRequest(
+            url: Uri.parse(
+              _oneSignalHelper.url!,
+            ),
+          ),
+        );
       }
     }
 
     _oneSignalHelper.addListener(_listenerOneSignal);
 
-    this.setState(() {
-      settings = widget.settings;
-    });
-
-    final List<Widget> _children = [];
-
-    print("widget.settings.floating.length");
-    print(widget.settings.floating.length);
+    // final List<Widget> _children = [];
 
     return WillPopScope(
       onWillPop: () async {
-        getCurrentKey().currentState.goBack();
+        return getCurrentKey().currentState!.goBack();
       },
       child: Container(
           decoration: BoxDecoration(color: HexColor("#f5f4f4")),
           padding: EdgeInsets.only(bottom: bottomPadding),
           child: Scaffold(
             key: _scaffoldKey,
-            appBar: _renderAppBar(context, settings),
-            drawer: (widget.settings.leftNavigationIcon.value == "icon_menu" ||
-                    widget.settings.rightNavigationIcon.value == "icon_menu")
+            appBar: _renderAppBar(context, settings) as PreferredSizeWidget?,
+            drawer: (widget.settings.leftNavigationIcon!.value == "icon_menu" ||
+                    widget.settings.rightNavigationIcon!.value == "icon_menu")
                 ? Drawer(
                     child: ListView(
                       padding: const EdgeInsets.all(0.0),
                       children: <Widget>[
                         DrawerHeader(
                             decoration: BoxDecoration(
-                                gradient: LinearGradient(colors: <Color>[
-                              themeProvider.isLightTheme
-                                  ? HexColor(settings.firstColor)
-                                  : themeProvider.darkTheme.primaryColor,
-                              themeProvider.isLightTheme
-                                  ? HexColor(settings.secondColor)
-                                  : themeProvider.darkTheme.primaryColor,
-                            ])),
+                              gradient: LinearGradient(
+                                colors: <Color>[
+                                  themeProvider.isLightTheme!
+                                      ? HexColor(settings.firstColor)
+                                      : themeProvider.darkTheme.primaryColor,
+                                  themeProvider.isLightTheme!
+                                      ? HexColor(settings.secondColor)
+                                      : themeProvider.darkTheme.primaryColor,
+                                ],
+                              ),
+                            ),
                             child: Container(
                               child: Column(
                                 crossAxisAlignment: CrossAxisAlignment.start,
@@ -343,19 +354,19 @@ class _HomeScreen extends State<HomeScreen>
                                     width: 70.0,
                                     height: 70.0,
                                     child: Image.network(
-                                      settings.logoHeaderUrl,
+                                      settings.logoHeaderUrl!,
                                     ),
                                   ),
                                   Padding(
                                     padding: EdgeInsets.only(top: 5),
-                                    child: Text(settings.title,
+                                    child: Text(settings.title!,
                                         overflow: TextOverflow.ellipsis,
                                         style: TextStyle(
                                             color: Colors.white, fontSize: 16)),
                                   ),
                                   Padding(
                                     padding: EdgeInsets.only(top: 5),
-                                    child: Text(settings.subTitle,
+                                    child: Text(settings.subTitle!,
                                         overflow: TextOverflow.ellipsis,
                                         style: TextStyle(
                                             color: Colors.white, fontSize: 14)),
@@ -365,41 +376,39 @@ class _HomeScreen extends State<HomeScreen>
                             )),
                         DrawerListTitle(
                             icon: Icons.home,
-                            text: I18n.current.home,
+                            text: I18n.current!.home,
                             onTap: () async {
                               if (widget.settings.tabNavigationEnable == "1") {
                                 if (goToWeb) {
-                                  setState(() {
-                                    goToWeb = false;
-                                  });
+                                  setState(() => goToWeb = false);
                                   Navigator.pop(context);
+                                  await Navigator.push(
+                                    context,
+                                    PageTransition(
+                                      type: pageTransitionAnimation(context),
+                                      child: WebScreen(widget.settings.url),
+                                    ),
+                                  );
 
-                                  final result = await Navigator.push(
-                                      context,
-                                      PageTransition(
-                                          type: PageTransitionType.rightToLeft,
-                                          child: WebScreen(widget.settings.url,
-                                              widget.settings)));
-
-                                  setState(() {
-                                    goToWeb = true;
-                                  });
+                                  setState(() => goToWeb = true);
                                 }
                               } else {
-                                key0.currentState._webViewController?.loadUrl(
-                                    urlRequest: URLRequest(
-                                        url: Uri.parse(settings.url)));
+                                key0.currentState!._webViewController?.loadUrl(
+                                  urlRequest: URLRequest(
+                                    url: Uri.parse(url),
+                                  ),
+                                );
 
                                 Navigator.pop(context);
                               }
                             }),
-                        _renderMenuDrawer(settings.menus, context),
+                        _renderMenuDrawer(settings.menus!, context),
                         Padding(
                           padding: EdgeInsets.fromLTRB(0, 20, 0, 20),
                           child: Divider(height: 1, color: Colors.grey[400]),
                         ),
-                        _renderPageDrawer(settings.pages, context),
-                        settings.pages.length != 0
+                        _renderPageDrawer(settings.pages!, context),
+                        settings.pages!.length != 0
                             ? Padding(
                                 padding: EdgeInsets.fromLTRB(0, 20, 0, 20),
                                 child:
@@ -408,67 +417,101 @@ class _HomeScreen extends State<HomeScreen>
                             : Container(height: 0),
                         DrawerListTitle(
                             icon: Icons.brightness_medium,
-                            text: themeProvider.isLightTheme
-                                ? I18n.current.darkMode
-                                : I18n.current.lightMode,
+                            text: themeProvider.isLightTheme!
+                                ? I18n.current!.darkMode
+                                : I18n.current!.lightMode,
                             onTap: () {
-                              if (themeProvider.isLightTheme) {
+                              if (themeProvider.isLightTheme!) {
                                 themeProvider.setDarkMode();
                               } else {
                                 themeProvider.setLightMode();
                               }
                             }),
                         DrawerListTitle(
-                            icon: Icons.translate,
-                            text: I18n.current.languages,
-                            onTap: () {
-                              Navigator.push(
-                                  context,
-                                  PageTransition(
-                                      type: PageTransitionType.rightToLeft,
-                                      child: LanguageScreen()));
-                            }),
+                          icon: Icons.translate,
+                          text: I18n.current!.languages,
+                          onTap: () {
+                            Navigator.pop(context);
+                            Navigator.push(
+                              context,
+                              PageTransition(
+                                type: pageTransitionAnimation(context),
+                                child: LanguageScreen(),
+                              ),
+                            );
+                          },
+                        ),
                         DrawerListTitle(
-                            icon: Icons.info,
-                            text: I18n.current.about,
-                            onTap: () {
-                              Navigator.pop(context);
-                              Navigator.push(
-                                  context,
-                                  PageTransition(
-                                      type: PageTransitionType.rightToLeft,
-                                      child: AboutScreen()));
-                            }),
+                          icon: Icons.museum_outlined,
+                          text: 'Change Mosque',
+                          onTap: () {
+                            Navigator.pop(context);
+                            Navigator.push(
+                              context,
+                              PageTransition(
+                                type: pageTransitionAnimation(context),
+                                child: MosqueSearchScreen(),
+                              ),
+                            );
+                          },
+                        ),
+                        DrawerListTitle(
+                          icon: Icons.info,
+                          text: I18n.current!.about,
+                          onTap: () {
+                            Navigator.pop(context);
+                            Navigator.push(
+                              context,
+                              PageTransition(
+                                type: pageTransitionAnimation(context),
+                                child: AboutScreen(),
+                              ),
+                            );
+                          },
+                        ),
                         DrawerListTitle(
                             icon: Icons.share,
-                            text: I18n.current.share,
+                            text: I18n.current!.share,
                             onTap: () {
-                              shareApp(context, settings.title, settings.share);
+                              shareApp(
+                                  context, settings.title, settings.share!);
                             }),
                         DrawerListTitle(
-                            icon: Icons.star,
-                            text: I18n.current.rate,
-                            onTap: () => LaunchReview.launch(
-                                androidAppId: settings.androidId,
-                                iOSAppId: settings.iosId)),
-                        (_packageInfo.version !=
-                                (Platform.isAndroid
-                                    ? settings.versionAndroid
-                                    : settings.versionIos))
-                            ? new Column(children: [
+                          icon: Icons.star,
+                          text: I18n.current!.rate,
+                          onTap: () => LaunchReview.launch(
+                            androidAppId: settings.androidId,
+                            iOSAppId: settings.iosId,
+                          ),
+                        ),
+                        FutureBuilder<PackageInfo>(
+                          future: PackageInfo.fromPlatform(),
+                          builder: (context, snapshot) {
+                            if (snapshot.data == null ||
+                                snapshot.data!.version ==
+                                    (Platform.isAndroid
+                                        ? settings.versionAndroid
+                                        : settings.versionIos))
+                              return SizedBox();
+                            return new Column(
+                              children: [
                                 Padding(
                                   padding: EdgeInsets.fromLTRB(0, 20, 0, 20),
                                   child: Divider(
                                       height: 1, color: Colors.grey[400]),
                                 ),
                                 DrawerListTitle(
-                                    icon: Icons.system_update,
-                                    text: I18n.current.update,
-                                    onTap: () => LaunchReview.launch(
-                                        androidAppId: settings.androidId,
-                                        iOSAppId: settings.iosId))
-                              ])
-                            : Container()
+                                  icon: Icons.system_update,
+                                  text: I18n.current!.update,
+                                  onTap: () => LaunchReview.launch(
+                                    androidAppId: settings.androidId,
+                                    iOSAppId: settings.iosId,
+                                  ),
+                                )
+                              ],
+                            );
+                          },
+                        ),
                       ],
                     ),
                   )
@@ -482,33 +525,28 @@ class _HomeScreen extends State<HomeScreen>
                         ? TabBarView(
                             controller: tabController,
                             physics: NeverScrollableScrollPhysics(),
-                            children: List.generate(widget.settings.tab.length,
-                                (index) {
-                              return WebViewScreen(
+                            children: List.generate(
+                              widget.settings.tab!.length,
+                              (index) => WebViewScreen(
                                   key: listKey[index],
-                                  path: widget.settings.tab[index].url,
+                                  path: widget.settings.tab![index].url,
                                   pos: index,
                                   settings: widget.settings,
-                                  stream: listStream[index].stream);
-                            }),
+                                  stream: listStream[index].stream),
+                            ),
                           )
-                        : TabBarView(
-                            controller: tabController,
-                            physics: NeverScrollableScrollPhysics(),
-                            children: List.generate(1, (index) {
-                              return WebViewScreen(
-                                  key: listKey[0],
-                                  path: widget.url,
-                                  pos: index,
-                                  settings: widget.settings,
-                                  stream: listStream[0].stream);
-                            }),
+                        : WebViewScreen(
+                            key: listKey[0],
+                            path: url,
+                            pos: 0,
+                            settings: widget.settings,
+                            stream: listStream[0].stream,
                           ),
                   ),
                   if (widget.settings.adBanner == "1" && _isBannerAdReady)
                     Container(
                       height: 50,
-                      child: AdWidget(ad: _bannerAd),
+                      child: AdWidget(ad: _bannerAd!),
                     )
                 ]),
               ],
@@ -530,17 +568,25 @@ class _HomeScreen extends State<HomeScreen>
                           ],
                         ),
                         height: 60.0,
-                        child: _buildTabItem(context)),
+                        child: _buildTabItem(context, settings)),
                   )
                 : Container(height: 0),
-            floatingActionButton: widget.settings.floating.length != 0
+            // floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
+            floatingActionButton: widget.settings.floating!.length != 0
                 ? SpeedDial(
                     icon: Icons.add,
                     backgroundColor: HexColor(widget.settings.firstColor),
                     foregroundColor: Colors.white,
                     children:
-                        _renderFloating(widget.settings.floating, context))
-                : null,
+                        _renderFloating(widget.settings.floating!, context),
+                  )
+                : Padding(
+                    padding: const EdgeInsets.only(bottom: 40),
+                    child: FloatingActionButton(
+                      onPressed: () => _scaffoldKey.currentState?.openDrawer(),
+                      child: Icon(Icons.menu),
+                    ),
+                  ),
             /*bottomNavigationBar: Container(
                 height: settings.adBanner == "1"
                     ? Platform.isAndroid
@@ -552,59 +598,68 @@ class _HomeScreen extends State<HomeScreen>
     );
   }
 
-  Widget _buildTabItem(context) {
+  Widget _buildTabItem(context, Settings settings) {
     var themeProvider = Provider.of<ThemeNotifier>(context);
     Color tabColor = HexColor(widget.settings.colorTab);
     Color unselectedColor = Colors.black26;
     return new TabBar(
-        onTap: (index) {
-          for (int i = 0; i < listStream.length; i++) {
-            listStream[i].add(index);
-          }
-        },
-        indicator: UnderlineTabIndicator(
-          borderSide: BorderSide(
-              color: themeProvider.isLightTheme
-                  ? tabColor
-                  : themeProvider.darkTheme.primaryColor,
-              width: 2.5),
-          //insets: EdgeInsets.symmetric(horizontal:16.0)
+      onTap: (index) {
+        for (int i = 0; i < listStream.length; i++) {
+          listStream[i].add(index);
+        }
+      },
+      indicator: UnderlineTabIndicator(
+        borderSide: BorderSide(
+          color: themeProvider.isLightTheme!
+              ? tabColor
+              : themeProvider.darkTheme.primaryColor,
+          width: 2.5,
         ),
-        controller: tabController,
-        labelColor: tabColor,
-        unselectedLabelColor: Colors.black26,
-        tabs: List.generate(widget.settings.tab.length, (index) {
+        //insets: EdgeInsets.symmetric(horizontal:16.0)
+      ),
+      controller: tabController,
+      labelColor: tabColor,
+      unselectedLabelColor: Colors.black26,
+      tabs: List.generate(
+        widget.settings.tab!.length,
+        (index) {
           return new Tab(
             child: new Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: <Widget>[
-                  new Image.network(settings.tab[index].icon_url,
-                      width: 25,
-                      height: 25,
+              mainAxisAlignment: MainAxisAlignment.center,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                new Image.network(settings.tab![index].icon_url!,
+                    width: 25,
+                    height: 25,
+                    color: _currentIndex == index
+                        ? themeProvider.isLightTheme!
+                            ? tabColor
+                            : themeProvider.darkTheme.primaryColor
+                        : unselectedColor),
+                new SizedBox(height: 5),
+                new Flexible(
+                  child: new Text(
+                    settings.tab![index].title!,
+                    overflow: TextOverflow.ellipsis,
+                    style: TextStyle(
+                      fontSize: widget.settings.tab!.length == 5 ? 8 : 10,
                       color: _currentIndex == index
-                          ? themeProvider.isLightTheme
+                          ? themeProvider.isLightTheme!
                               ? tabColor
                               : themeProvider.darkTheme.primaryColor
-                          : unselectedColor),
-                  new SizedBox(height: 5),
-                  new Flexible(
-                      child: new Text(settings.tab[index].title,
-                          overflow: TextOverflow.ellipsis,
-                          style: TextStyle(
-                              fontSize:
-                                  widget.settings.tab.length == 5 ? 8 : 10,
-                              color: _currentIndex == index
-                                  ? themeProvider.isLightTheme
-                                      ? tabColor
-                                      : themeProvider.darkTheme.primaryColor
-                                  : unselectedColor))),
-                ]),
+                          : unselectedColor,
+                    ),
+                  ),
+                ),
+              ],
+            ),
           );
-        }));
+        },
+      ),
+    );
   }
 
-  Widget _offline(bottomPadding) {
+  Widget _offline(bottomPadding, Settings settings) {
     return WillPopScope(
       onWillPop: () async {
         return _onBackPressed(context);
@@ -632,7 +687,7 @@ class _HomeScreen extends State<HomeScreen>
                           )),
                       SizedBox(height: 40),
                       Text(
-                        I18n.current.whoops,
+                        I18n.current!.whoops,
                         style: TextStyle(
                             color: Colors.black45,
                             fontSize: 40.0,
@@ -640,14 +695,14 @@ class _HomeScreen extends State<HomeScreen>
                       ),
                       SizedBox(height: 20),
                       Text(
-                        I18n.current.noInternet,
+                        I18n.current!.noInternet,
                         style: TextStyle(color: Colors.black87, fontSize: 15.0),
                       ),
                       SizedBox(height: 5),
                       SizedBox(height: 60),
                       RaisedGradientButton(
                           child: Text(
-                            I18n.current.tryAgain,
+                            I18n.current!.tryAgain,
                             style: TextStyle(
                                 color: Colors.white,
                                 fontSize: 18.0,
@@ -682,7 +737,7 @@ class _HomeScreen extends State<HomeScreen>
   }
    */
 
-  int parseInt(dynamic value) {
+  int? parseInt(dynamic value) {
     if (value == null) return null;
     if (value is int) return value;
 
@@ -693,25 +748,25 @@ class _HomeScreen extends State<HomeScreen>
       PositionOptions positionOptions) async {
     PositionResponse positionResponse = PositionResponse();
 
-    int timeout = 30000;
-    if (positionOptions.timeout > 0) timeout = positionOptions.timeout;
+    int? timeout = 30000;
+    if (positionOptions.timeout! > 0) timeout = positionOptions.timeout;
 
     try {
-      Geolocator geolocator = Geolocator()..forceAndroidLocationManager = true;
-      GeolocationStatus geolocationStatus =
-          await geolocator.checkGeolocationPermissionStatus();
+      // Geolocator geolocator = Geolocator()..forceAndroidLocationManager = true;
+      LocationPermission geolocationStatus =
+          await GeolocatorPlatform.instance.requestPermission();
 
-      if (geolocationStatus == GeolocationStatus.denied ||
-          geolocationStatus == GeolocationStatus.granted ||
-          geolocationStatus == GeolocationStatus.restricted ||
-          geolocationStatus == GeolocationStatus.unknown) {
+      if (geolocationStatus == LocationPermission.always ||
+          geolocationStatus == LocationPermission.whileInUse) {
         positionResponse.position = await Future.any([
-          geolocator.getCurrentPosition(
-              desiredAccuracy: (positionOptions.enableHighAccuracy
-                  ? LocationAccuracy.best
-                  : LocationAccuracy.medium)),
-          Future.delayed(Duration(milliseconds: timeout), () {
-            if (positionOptions.timeout > 0) positionResponse.timedOut = true;
+          GeolocatorPlatform.instance.getCurrentPosition(
+            locationSettings: LocationSettings(
+                accuracy: (positionOptions.enableHighAccuracy
+                    ? LocationAccuracy.best
+                    : LocationAccuracy.medium)),
+          ),
+          Future.delayed(Duration(milliseconds: timeout!), () {
+            if (positionOptions.timeout! > 0) positionResponse.timedOut = true;
             return;
           })
         ]);
@@ -739,12 +794,18 @@ class _HomeScreen extends State<HomeScreen>
     return positionResponse;
   }
 
-  Widget _renderMenuDrawer(List<Menu> menus, context) {
+  pageTransitionAnimation(BuildContext context) {
+    return Directionality.of(context) == TextDirection.ltr
+        ? PageTransitionType.leftToRight
+        : PageTransitionType.rightToLeft;
+  }
+
+  Widget _renderMenuDrawer(List<Menu> menus, BuildContext context) {
     return new Column(
       children: menus
           .map(
             (Menu menu) => DrawerListTitle(
-                icon_url: menu.iconUrl,
+                iconUrl: menu.iconUrl,
                 text: menu.title,
                 onTap: () async {
                   if (widget.settings.tabNavigationEnable == "1") {
@@ -753,18 +814,20 @@ class _HomeScreen extends State<HomeScreen>
                         goToWeb = false;
                       });
                       final result = await Navigator.push(
-                          context,
-                          PageTransition(
-                              type: PageTransitionType.rightToLeft,
-                              child: WebScreen(menu.url, widget.settings)));
+                        context,
+                        PageTransition(
+                          type: pageTransitionAnimation(context),
+                          child: WebScreen(menu.url),
+                        ),
+                      );
 
                       setState(() {
                         goToWeb = true;
                       });
                     }
                   } else {
-                    key0.currentState._webViewController?.loadUrl(
-                        urlRequest: URLRequest(url: Uri.parse(menu.url)));
+                    key0.currentState!._webViewController?.loadUrl(
+                        urlRequest: URLRequest(url: Uri.parse(menu.url!)));
 
                     Navigator.pop(context);
                   }
@@ -780,7 +843,7 @@ class _HomeScreen extends State<HomeScreen>
           (Floating floating) => SpeedDialChild(
               child: Container(
                 padding: EdgeInsets.all(13.0),
-                child: Image.network(floating.iconUrl,
+                child: Image.network(floating.iconUrl!,
                     width: 15, height: 15, color: HexColor(floating.iconColor)),
               ),
               label: floating.title,
@@ -794,18 +857,21 @@ class _HomeScreen extends State<HomeScreen>
                       goToWeb = false;
                     });
                     final result = await Navigator.push(
-                        context,
-                        PageTransition(
-                            type: PageTransitionType.rightToLeft,
-                            child: WebScreen(floating.url, widget.settings)));
+                      context,
+                      PageTransition(
+                        type: pageTransitionAnimation(context),
+                        child: WebScreen(floating.url),
+                      ),
+                    );
 
-                    setState(() {
-                      goToWeb = true;
-                    });
+                    setState(() => goToWeb = true);
                   }
                 } else {
-                  key0.currentState._webViewController?.loadUrl(
-                      urlRequest: URLRequest(url: Uri.parse(floating.url)));
+                  key0.currentState!._webViewController?.loadUrl(
+                    urlRequest: URLRequest(
+                      url: Uri.parse(floating.url!),
+                    ),
+                  );
 
                   //Navigator.pop(context);
                 }
@@ -815,21 +881,21 @@ class _HomeScreen extends State<HomeScreen>
   }
 
   Widget _renderPageDrawer(List<Page> pages, context) {
+    print(pages.map((e) => e.icon));
     return new Column(
       children: pages
-          .map(
-            (Page page) => DrawerListTitle(
-                icon_url: page.iconUrl,
-                text: page.title,
-                onTap: () {
-                  Navigator.pop(context);
-                  Navigator.push(
-                      context,
-                      PageTransition(
-                          type: PageTransitionType.rightToLeft,
-                          child: PageScreen(page, widget.settings)));
-                }),
-          )
+          .map((Page page) => DrawerListTitle(
+              forceThemeColor: true,
+              iconUrl: page.iconUrl,
+              text: page.title,
+              onTap: () {
+                Navigator.pop(context);
+                Navigator.push(
+                    context,
+                    PageTransition(
+                        type: pageTransitionAnimation(context),
+                        child: PageScreen(page)));
+              }))
           .toList(),
     );
   }
@@ -844,7 +910,7 @@ class _HomeScreen extends State<HomeScreen>
                 children: <Widget>[
                   _renderMenuIcon(
                       context,
-                      widget.settings.leftNavigationIcon,
+                      widget.settings.leftNavigationIcon!,
                       widget.settings.rightNavigationIcon,
                       widget.settings.navigatinBarStyle,
                       widget.settings,
@@ -854,7 +920,7 @@ class _HomeScreen extends State<HomeScreen>
                   Row(
                       children: _renderMenuIconList(
                           context,
-                          widget.settings.rightNavigationIconList,
+                          widget.settings.rightNavigationIconList!,
                           widget.settings.leftNavigationIcon,
                           widget.settings.navigatinBarStyle,
                           widget.settings,
@@ -867,10 +933,10 @@ class _HomeScreen extends State<HomeScreen>
                   begin: Alignment.centerLeft,
                   end: Alignment.centerRight,
                   colors: <Color>[
-                    themeProvider.isLightTheme
+                    themeProvider.isLightTheme!
                         ? HexColor(settings.firstColor)
                         : themeProvider.darkTheme.primaryColor,
-                    themeProvider.isLightTheme
+                    themeProvider.isLightTheme!
                         ? HexColor(settings.secondColor)
                         : themeProvider.darkTheme.primaryColor,
                   ],
@@ -884,7 +950,7 @@ class _HomeScreen extends State<HomeScreen>
             ));
   }
 
-  Widget _renderTitle(String type, Settings settings) {
+  Widget _renderTitle(String? type, Settings settings) {
     var direction = MainAxisAlignment.start;
 
     switch (type) {
@@ -909,7 +975,7 @@ class _HomeScreen extends State<HomeScreen>
             child: Container(
               child: settings.typeHeader == "text"
                   ? Text(
-                      settings.title,
+                      settings.title!,
                       overflow: TextOverflow.ellipsis,
                       style: TextStyle(
                           color: Colors.white,
@@ -917,7 +983,7 @@ class _HomeScreen extends State<HomeScreen>
                           fontWeight: FontWeight.bold),
                     )
                   : settings.typeHeader == "image"
-                      ? Image.network(settings.logoHeaderUrl, height: 40)
+                      ? Image.network(settings.logoHeaderUrl!, height: 40)
                       : Container(),
             ),
           )
@@ -929,8 +995,8 @@ class _HomeScreen extends State<HomeScreen>
   Widget _renderMenuIcon(
       BuildContext context,
       NavigationIcon navigationIcon,
-      NavigationIcon navigationOtherIcon,
-      String navigatinBarStyle,
+      NavigationIcon? navigationOtherIcon,
+      String? navigatinBarStyle,
       Settings settings,
       String direction) {
     return navigationIcon.value != "icon_empty"
@@ -945,10 +1011,10 @@ class _HomeScreen extends State<HomeScreen>
                       icon: Transform(
                           alignment: Alignment.center,
                           transform: Matrix4.rotationY(math.pi *
-                              (I18n.current.textDirection == TextDirection.ltr
+                              (I18n.current!.textDirection == TextDirection.ltr
                                   ? 2
                                   : 1)),
-                          child: new Image.network(navigationIcon.iconUrl,
+                          child: new Image.network(navigationIcon.iconUrl!,
                               height: 25, width: 25, color: Colors.white)
                           /*Image.asset(
                               UIImages.imageDir +
@@ -965,7 +1031,7 @@ class _HomeScreen extends State<HomeScreen>
                     ),
                     Container(
                       width: (navigatinBarStyle == "center" &&
-                              navigationOtherIcon.value == "icon_back_forward")
+                              navigationOtherIcon!.value == "icon_back_forward")
                           ? 50
                           : 0,
                     )
@@ -978,7 +1044,8 @@ class _HomeScreen extends State<HomeScreen>
                         icon: Transform(
                             alignment: Alignment.center,
                             transform: Matrix4.rotationY(math.pi *
-                                (I18n.current.textDirection == TextDirection.ltr
+                                (I18n.current!.textDirection ==
+                                        TextDirection.ltr
                                     ? 2
                                     : 1)),
                             child: Image.asset(
@@ -990,29 +1057,29 @@ class _HomeScreen extends State<HomeScreen>
                           switch (_currentIndex) {
                             case 0:
                               {
-                                key0.currentState._webViewController?.goBack();
+                                key0.currentState!._webViewController?.goBack();
                               }
                               break;
 
                             case 1:
                               {
-                                key1.currentState._webViewController?.goBack();
+                                key1.currentState!._webViewController?.goBack();
                               }
                               break;
 
                             case 2:
                               {
-                                key2.currentState._webViewController?.goBack();
+                                key2.currentState!._webViewController?.goBack();
                               }
                               break;
                             case 3:
                               {
-                                key3.currentState._webViewController?.goBack();
+                                key3.currentState!._webViewController?.goBack();
                               }
                               break;
                             case 4:
                               {
-                                key4.currentState._webViewController?.goBack();
+                                key4.currentState!._webViewController?.goBack();
                               }
                               break;
                             default:
@@ -1028,7 +1095,8 @@ class _HomeScreen extends State<HomeScreen>
                         icon: Transform(
                             alignment: Alignment.center,
                             transform: Matrix4.rotationY(math.pi *
-                                (I18n.current.textDirection == TextDirection.ltr
+                                (I18n.current!.textDirection ==
+                                        TextDirection.ltr
                                     ? 2
                                     : 1)),
                             child: Image.asset(
@@ -1038,7 +1106,7 @@ class _HomeScreen extends State<HomeScreen>
                                 color: Colors.white)),
                         onPressed: () {
                           getCurrentKey()
-                              .currentState
+                              .currentState!
                               ._webViewController
                               ?.goForward();
                         },
@@ -1054,8 +1122,8 @@ class _HomeScreen extends State<HomeScreen>
   List<Widget> _renderMenuIconList(
       BuildContext context,
       List<NavigationIcon> navigationIcon,
-      NavigationIcon navigationOtherIcon,
-      String navigatinBarStyle,
+      NavigationIcon? navigationOtherIcon,
+      String? navigatinBarStyle,
       Settings settings,
       String direction) {
     return navigationIcon
@@ -1074,9 +1142,12 @@ class _HomeScreen extends State<HomeScreen>
 
   Future<bool> _onBackPressed(context) async {
     try {
-      if (getCurrentKey().currentState._webViewController != null) {
-        if (await getCurrentKey().currentState._webViewController.canGoBack()) {
-          getCurrentKey().currentState._webViewController.goBack();
+      if (getCurrentKey().currentState!._webViewController != null) {
+        if (await getCurrentKey()
+            .currentState!
+            ._webViewController!
+            .canGoBack()) {
+          getCurrentKey().currentState!._webViewController!.goBack();
           return false;
         } else {
           _showDialog(context);
@@ -1085,23 +1156,25 @@ class _HomeScreen extends State<HomeScreen>
     } catch (e) {
       _showDialog(context);
     }
+
+    return true;
   }
 
   _showDialog(context) {
     return showDialog(
       context: context,
       builder: (context) => new AlertDialog(
-        title: new Text(I18n.current.closeApp),
-        content: new Text(I18n.current.sureCloseApp),
+        title: new Text(I18n.current!.closeApp),
+        content: new Text(I18n.current!.sureCloseApp),
         actions: <Widget>[
           new FlatButton(
             onPressed: () => Navigator.of(context).pop(false),
-            child: new Text(I18n.current.cancel),
+            child: new Text(I18n.current!.cancel),
           ),
           SizedBox(height: 16),
           new FlatButton(
             onPressed: () => exit(0),
-            child: new Text(I18n.current.ok),
+            child: new Text(I18n.current!.ok),
           ),
         ],
       ),
@@ -1121,20 +1194,20 @@ class _HomeScreen extends State<HomeScreen>
           final result = await Navigator.push(
               context,
               PageTransition(
-                  type: PageTransitionType.rightToLeft,
-                  child: WebScreen(navigationIcon.url, widget.settings)));
+                  type: pageTransitionAnimation(context),
+                  child: WebScreen(navigationIcon.url)));
           setState(() {
             goToWeb = false;
           });
         }
       } else {
-        key0.currentState._webViewController?.loadUrl(
-            urlRequest: URLRequest(url: Uri.parse(navigationIcon.url)));
+        key0.currentState!._webViewController?.loadUrl(
+            urlRequest: URLRequest(url: Uri.parse(navigationIcon.url!)));
       }
     } else {
       switch (navigationIcon.value) {
         case "icon_menu":
-          _HomeScreen._scaffoldKey.currentState.openDrawer();
+          _HomeScreen._scaffoldKey.currentState!.openDrawer();
           break;
         case "icon_home":
           if (widget.settings.tabNavigationEnable == "1") {
@@ -1145,28 +1218,28 @@ class _HomeScreen extends State<HomeScreen>
               final result = await Navigator.push(
                   context,
                   PageTransition(
-                      type: PageTransitionType.rightToLeft,
-                      child: WebScreen(settings.url, widget.settings)));
+                      type: pageTransitionAnimation(context),
+                      child: WebScreen(settings.url)));
               setState(() {
                 goToWeb = true;
               });
             }
           } else {
-            key0.currentState._webViewController
-                ?.loadUrl(urlRequest: URLRequest(url: Uri.parse(settings.url)));
+            key0.currentState!._webViewController?.loadUrl(
+                urlRequest: URLRequest(url: Uri.parse(settings.url!)));
           }
           break;
         case "icon_reload":
-          getCurrentKey().currentState._webViewController?.reload();
+          getCurrentKey().currentState!._webViewController?.reload();
           break;
         case "icon_share":
-          shareApp(context, settings.title, settings.share);
+          shareApp(context, settings.title, settings.share!);
           break;
         case "icon_back":
-          getCurrentKey().currentState._webViewController?.goBack();
+          getCurrentKey().currentState!._webViewController?.goBack();
           break;
         case "icon_forward":
-          getCurrentKey().currentState._webViewController?.goForward();
+          getCurrentKey().currentState!._webViewController?.goForward();
           break;
         case "icon_exit":
           _showDialog(context);
@@ -1193,14 +1266,14 @@ class _HomeScreen extends State<HomeScreen>
       if (!mounted) return;
       print("qrCode" + qrCode);
       setState(() {
-        key0.currentState.isLoading = true;
+        key0.currentState!.isLoading = true;
       });
 
       final response = await http.head(Uri.parse(qrCode));
 
       if (response.statusCode == 200) {
         setState(() {
-          key0.currentState.isLoading = false;
+          key0.currentState!.isLoading = false;
         });
         if (widget.settings.tabNavigationEnable == "1") {
           if (goToWeb) {
@@ -1210,32 +1283,32 @@ class _HomeScreen extends State<HomeScreen>
             final result = await Navigator.push(
                 context,
                 PageTransition(
-                    type: PageTransitionType.rightToLeft,
-                    child: WebScreen(qrCode, widget.settings)));
+                    type: pageTransitionAnimation(context),
+                    child: WebScreen(qrCode)));
 
             setState(() {
               goToWeb = true;
             });
           }
         } else {
-          key0.currentState._webViewController
+          key0.currentState!._webViewController
               ?.loadUrl(urlRequest: URLRequest(url: Uri.parse(qrCode)));
         }
       } else {
         setState(() {
-          key0.currentState.isLoading = false;
+          key0.currentState!.isLoading = false;
         });
       }
     } finally {
       setState(() {
-        key0.currentState.isLoading = false;
+        key0.currentState!.isLoading = false;
       });
       print('Failed to get platform version.');
     }
   }
 
-  shareApp(BuildContext context, String text, String share) {
-    final RenderBox box = context.findRenderObject();
+  shareApp(BuildContext context, String? text, String share) {
+    final RenderBox box = context.findRenderObject() as RenderBox;
     Share.share(share,
         subject: text,
         sharePositionOrigin: box.localToGlobal(Offset.zero) & box.size);
@@ -1280,22 +1353,22 @@ class _HomeScreen extends State<HomeScreen>
 }
 
 class WebViewScreen extends StatefulWidget {
-  final GlobalKey webKey;
-  final String path;
-  final int pos;
-  final Settings settings;
-  final InAppWebViewController webViewController;
-  final Stream<int> stream;
+  final GlobalKey? webKey;
+  final String? path;
+  final int? pos;
+  final Settings? settings;
+  final InAppWebViewController? webViewController;
+  final Stream<int>? stream;
 
-  WebViewScreen(
-      {Key key,
-      this.path,
-      this.webKey,
-      this.pos,
-      this.settings,
-      this.webViewController,
-      this.stream})
-      : super(key: key);
+  WebViewScreen({
+    Key? key,
+    this.path,
+    this.webKey,
+    this.pos,
+    this.settings,
+    this.webViewController,
+    this.stream,
+  }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() {
@@ -1308,13 +1381,13 @@ class _WebViewScreen extends State<WebViewScreen>
   @override
   bool get wantKeepAlive => true;
 
-  InAppWebViewController _webViewController;
+  InAppWebViewController? _webViewController;
   String url = "";
-  PullToRefreshController pullToRefreshController;
+  PullToRefreshController? pullToRefreshController;
   static GlobalKey<ScaffoldState> _scaffoldKey = new GlobalKey<ScaffoldState>();
 
   List<StreamSubscription<Position>> webViewGPSPositionStreams = [];
-  bool isLoading;
+  late bool isLoading;
 
   final Set<Factory<OneSequenceGestureRecognizer>> _gSet = [
     Factory<VerticalDragGestureRecognizer>(
@@ -1325,7 +1398,7 @@ class _WebViewScreen extends State<WebViewScreen>
 
   @override
   void initState() {
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
     super.initState();
     isLoading = true;
 
@@ -1363,6 +1436,18 @@ class _WebViewScreen extends State<WebViewScreen>
   }
 
   @override
+  void didUpdateWidget(covariant WebViewScreen oldWidget) {
+    if (oldWidget.path != widget.path) {
+      _webViewController!.loadUrl(
+        urlRequest: URLRequest(
+          url: Uri.parse(widget.path!),
+        ),
+      );
+    }
+    super.didUpdateWidget(oldWidget);
+  }
+
+  @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
     /*if (_webViewController != null) {
       if (state == AppLifecycleState.paused) {
@@ -1381,7 +1466,7 @@ class _WebViewScreen extends State<WebViewScreen>
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     webViewGPSPositionStreams.forEach(
         (StreamSubscription<Position> _flutterGeolocationStream) =>
             _flutterGeolocationStream.cancel());
@@ -1399,24 +1484,25 @@ class _WebViewScreen extends State<WebViewScreen>
           Expanded(
             child: InAppWebView(
               // contextMenu: contextMenu,
-              initialUrlRequest: URLRequest(url: Uri.parse(widget.path)),
+              initialUrlRequest: URLRequest(url: Uri.parse(widget.path!)),
               gestureRecognizers: _gSet,
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                      supportZoom: false,
-                      useShouldOverrideUrlLoading: true,
-                      useOnDownloadStart: true,
-                      mediaPlaybackRequiresUserGesture: false,
-                      userAgent: Platform.isAndroid
-                          ? widget.settings.userAgent.valueAndroid
-                          : widget.settings.userAgent.valueIOS),
+                    supportZoom: false,
+                    useShouldOverrideUrlLoading: true,
+                    useOnDownloadStart: true,
+                    mediaPlaybackRequiresUserGesture: false,
+                    userAgent: Platform.isAndroid
+                        ? widget.settings!.userAgent!.valueAndroid!
+                        : widget.settings!.userAgent!.valueIOS!,
+                  ),
                   android: AndroidInAppWebViewOptions(
                     useHybridComposition: true,
                   ),
                   ios: IOSInAppWebViewOptions(
                     allowsInlineMediaPlayback: true,
                   )),
-              pullToRefreshController: widget.settings.pullRefresh == "1"
+              pullToRefreshController: widget.settings!.pullRefresh == "1"
                   ? pullToRefreshController
                   : null,
               onWebViewCreated: (InAppWebViewController controller) {
@@ -1438,7 +1524,8 @@ class _WebViewScreen extends State<WebViewScreen>
                       switch (action) {
                         case "clearWatch":
                           _geolocationClearWatch(parseInt(
-                              geolocationData['flutterGeolocationIndex'] ?? 0));
+                              geolocationData['flutterGeolocationIndex'] ??
+                                  0)!);
                           break;
 
                         case "getCurrentPosition":
@@ -1454,7 +1541,7 @@ class _WebViewScreen extends State<WebViewScreen>
                           _geolocationWatchPosition(
                               parseInt(
                                   geolocationData['flutterGeolocationIndex'] ??
-                                      0),
+                                      0)!,
                               PositionOptions()
                                   .from(geolocationData['option'] ?? null));
                           break;
@@ -1470,7 +1557,7 @@ class _WebViewScreen extends State<WebViewScreen>
                     action: PermissionRequestResponseAction.GRANT);
               },
               shouldOverrideUrlLoading: (controller, navigationAction) async {
-                var uri = navigationAction.request.url;
+                var uri = navigationAction.request.url!;
                 print("uri.scheme");
                 print(uri.scheme);
                 if (Platform.isAndroid && ["intent"].contains(uri.scheme)) {
@@ -1516,7 +1603,7 @@ class _WebViewScreen extends State<WebViewScreen>
                 });
               },
               onLoadStop: (controller, url) async {
-                pullToRefreshController.endRefreshing();
+                pullToRefreshController!.endRefreshing();
                 Future.delayed(const Duration(milliseconds: 500), () {
                   _geolocationAlertFix();
                 });
@@ -1536,11 +1623,11 @@ class _WebViewScreen extends State<WebViewScreen>
                 }
               },
               onLoadError: (controller, url, code, message) {
-                pullToRefreshController.endRefreshing();
+                pullToRefreshController!.endRefreshing();
               },
               onProgressChanged: (controller, progress) {
                 if (progress == 100) {
-                  pullToRefreshController.endRefreshing();
+                  pullToRefreshController!.endRefreshing();
                 }
               },
               onUpdateVisitedHistory: (controller, url, androidIsReload) {
@@ -1554,23 +1641,23 @@ class _WebViewScreen extends State<WebViewScreen>
             ),
           )
         ]),
-        (isLoading && widget.settings.loader != "empty")
+        (isLoading && widget.settings!.loader != "empty")
             ? Positioned(
                 top: 0,
                 bottom: 0,
                 right: 0,
                 left: 0,
                 child: Loader(
-                    type: widget.settings.loader,
-                    color: themeProvider.isLightTheme
-                        ? HexColor(widget.settings.loaderColor)
+                    type: widget.settings!.loader,
+                    color: themeProvider.isLightTheme!
+                        ? HexColor(widget.settings!.loaderColor)
                         : themeProvider.darkTheme.primaryColor))
             : Container()
       ],
     );
   }
 
-  int parseInt(dynamic value) {
+  int? parseInt(dynamic value) {
     if (value == null) return null;
     if (value is int) return value;
 
@@ -1581,25 +1668,22 @@ class _WebViewScreen extends State<WebViewScreen>
       PositionOptions positionOptions) async {
     PositionResponse positionResponse = PositionResponse();
 
-    int timeout = 30000;
-    if (positionOptions.timeout > 0) timeout = positionOptions.timeout;
+    int? timeout = 30000;
+    if (positionOptions.timeout! > 0) timeout = positionOptions.timeout;
 
     try {
-      Geolocator geolocator = Geolocator()..forceAndroidLocationManager = true;
-      GeolocationStatus geolocationStatus =
-          await geolocator.checkGeolocationPermissionStatus();
+      LocationPermission geolocationStatus =
+          await GeolocatorPlatform.instance.requestPermission();
 
-      if (geolocationStatus == GeolocationStatus.denied ||
-          geolocationStatus == GeolocationStatus.granted ||
-          geolocationStatus == GeolocationStatus.restricted ||
-          geolocationStatus == GeolocationStatus.unknown) {
+      if (geolocationStatus == LocationPermission.always ||
+          geolocationStatus == LocationPermission.whileInUse) {
         positionResponse.position = await Future.any([
-          geolocator.getCurrentPosition(
+          Geolocator.getCurrentPosition(
               desiredAccuracy: (positionOptions.enableHighAccuracy
                   ? LocationAccuracy.best
                   : LocationAccuracy.medium)),
-          Future.delayed(Duration(milliseconds: timeout), () {
-            if (positionOptions.timeout > 0) positionResponse.timedOut = true;
+          Future.delayed(Duration(milliseconds: timeout!), () {
+            if (positionOptions.timeout! > 0) positionResponse.timedOut = true;
             return;
           })
         ]);
@@ -1655,9 +1739,9 @@ class _WebViewScreen extends State<WebViewScreen>
       setTimeout(function(){ _flutterGeolocationAlertFix(); }, 100);
     ''';
 
-    _webViewController.evaluateJavascript(source: javascript);
+    _webViewController!.evaluateJavascript(source: javascript);
 
-    _webViewController.evaluateJavascript(source: """
+    _webViewController!.evaluateJavascript(source: """
       function _flutterGeolocationAlertFix() {
         navigator.geolocation = {};
         navigator.geolocation.clearWatch = function(watchId) {
@@ -1694,7 +1778,7 @@ class _WebViewScreen extends State<WebViewScreen>
 
   void _geolocationClearWatch(int flutterGeolocationIndex) {
     // Stop gps position stream
-    webViewGPSPositionStreams[flutterGeolocationIndex]?.cancel();
+    webViewGPSPositionStreams[flutterGeolocationIndex].cancel();
 
     // remove watcher from list
     webViewGPSPositionStreams.remove(flutterGeolocationIndex);
@@ -1713,11 +1797,11 @@ class _WebViewScreen extends State<WebViewScreen>
       _flutterGeolocationResponse();
     ''';
 
-    _webViewController.evaluateJavascript(source: javascript);
+    _webViewController!.evaluateJavascript(source: javascript);
   }
 
   void _geolocationGetCurrentPosition(
-      int flutterGeolocationIndex, PositionOptions positionOptions) async {
+      int? flutterGeolocationIndex, PositionOptions positionOptions) async {
     PositionResponse positionResponse =
         await getCurrentPosition(positionOptions);
 
@@ -1726,7 +1810,7 @@ class _WebViewScreen extends State<WebViewScreen>
   }
 
   void _geolocationResponse(
-      int flutterGeolocationIndex,
+      int? flutterGeolocationIndex,
       PositionOptions positionOptions,
       PositionResponse positionResponse,
       bool watcher) {
@@ -1738,25 +1822,25 @@ class _WebViewScreen extends State<WebViewScreen>
           ''']({
             coords: {
               accuracy: ''' +
-          positionResponse.position.accuracy.toString() +
+          positionResponse.position!.accuracy.toString() +
           ''',
               altitude: ''' +
-          positionResponse.position.altitude.toString() +
+          positionResponse.position!.altitude.toString() +
           ''',
               altitudeAccuracy: null,
               heading: null,
               latitude: ''' +
-          positionResponse.position.latitude.toString() +
+          positionResponse.position!.latitude.toString() +
           ''',
               longitude: ''' +
-          positionResponse.position.longitude.toString() +
+          positionResponse.position!.longitude.toString() +
           ''',
               speed: ''' +
-          positionResponse.position.speed.toString() +
+          positionResponse.position!.speed.toString() +
           '''
             },
             timestamp: ''' +
-          positionResponse.position.timestamp.millisecondsSinceEpoch
+          positionResponse.position!.timestamp!.millisecondsSinceEpoch
               .toString() +
           '''
           });''' +
@@ -1776,7 +1860,7 @@ class _WebViewScreen extends State<WebViewScreen>
         _flutterGeolocationResponse();
       ''';
 
-      _webViewController.evaluateJavascript(source: javascript);
+      _webViewController!.evaluateJavascript(source: javascript);
     } else {
       // TODO: Return correct error code
       String javascript = '''
@@ -1808,23 +1892,24 @@ class _WebViewScreen extends State<WebViewScreen>
         _flutterGeolocationResponse();
       ''';
 
-      _webViewController.evaluateJavascript(source: javascript);
+      _webViewController!.evaluateJavascript(source: javascript);
     }
   }
 
   void _geolocationWatchPosition(
-      int flutterGeolocationIndex, PositionOptions positionOptions) {
+    int flutterGeolocationIndex,
+    PositionOptions positionOptions,
+  ) {
     // init new strem
-    var geolocator = Geolocator();
-    var locationOptions = LocationOptions(
+    var locationOptions = LocationSettings(
         accuracy: (positionOptions.enableHighAccuracy
             ? LocationAccuracy.best
             : LocationAccuracy.medium),
         distanceFilter: 10);
 
-    webViewGPSPositionStreams[flutterGeolocationIndex] = geolocator
-        .getPositionStream(locationOptions)
-        .listen((Position position) {
+    webViewGPSPositionStreams[flutterGeolocationIndex] =
+        Geolocator.getPositionStream(locationSettings: locationOptions)
+            .listen((Position position) {
       // Send data to each warcher
       PositionResponse positionResponse = PositionResponse()
         ..position = position;
@@ -1835,8 +1920,8 @@ class _WebViewScreen extends State<WebViewScreen>
 
   Future<bool> goBack() async {
     if (_webViewController != null) {
-      if (await _webViewController.canGoBack()) {
-        _webViewController.goBack();
+      if (await _webViewController!.canGoBack()) {
+        _webViewController!.goBack();
         return false;
       } else {
         return showDialog(
@@ -1857,7 +1942,7 @@ class _WebViewScreen extends State<WebViewScreen>
                   ),
                 ],
               ),
-            ) ??
+            ) as FutureOr<bool>? ??
             false;
       }
     }
