@@ -1,22 +1,20 @@
 import 'dart:io';
-import 'dart:ui';
 
 import 'package:flutter/material.dart' hide Page;
-import 'package:flutter/rendering.dart';
+import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:flyweb/src/elements/Loader.dart';
 import 'package:flyweb/src/helpers/HexColor.dart';
 import 'package:flyweb/src/models/page.dart';
 import 'package:flyweb/src/models/settings.dart';
+import 'package:flyweb/src/services/settings_manager.dart';
 import 'package:flyweb/src/services/theme_manager.dart';
 import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 
 class PageScreen extends StatefulWidget {
   final Page page;
-  final Settings settings;
 
-  const PageScreen(this.page, this.settings);
+  const PageScreen(this.page);
 
   @override
   State<StatefulWidget> createState() {
@@ -26,8 +24,8 @@ class PageScreen extends StatefulWidget {
 }
 
 class _PageScreen extends State<PageScreen> {
-  InAppWebViewController _webViewController;
-  bool isLoading;
+  InAppWebViewController? _webViewController;
+  late bool isLoading;
 
   @override
   void initState() {
@@ -40,13 +38,15 @@ class _PageScreen extends State<PageScreen> {
     // TODO: implement build
 
     var themeProvider = Provider.of<ThemeNotifier>(context);
+    final settingsManager = Provider.of<SettingsManager>(context);
+    final settings = settingsManager.settings;
 
     return Scaffold(
-        appBar: _renderAppBar(context, widget.settings, widget.page),
+        appBar: _renderAppBar(context, settings, widget.page),
         body: Stack(fit: StackFit.expand, children: [
           InAppWebView(
             // contextMenu: contextMenu,
-            initialUrlRequest: URLRequest(url: Uri.parse(widget.page.url)),
+            initialUrlRequest: URLRequest(url: Uri.parse(widget.page.url!)),
             initialOptions: InAppWebViewGroupOptions(
                 crossPlatform: InAppWebViewOptions(
                     supportZoom: false,
@@ -54,8 +54,8 @@ class _PageScreen extends State<PageScreen> {
                     useOnDownloadStart: true,
                     mediaPlaybackRequiresUserGesture: false,
                     userAgent: Platform.isAndroid
-                        ? widget.settings.userAgent.valueAndroid
-                        : widget.settings.userAgent.valueIOS),
+                        ? settings.userAgent!.valueAndroid!
+                        : settings.userAgent!.valueIOS!),
                 android: AndroidInAppWebViewOptions(
                   useHybridComposition: true,
                 ),
@@ -104,27 +104,27 @@ class _PageScreen extends State<PageScreen> {
               print(consoleMessage);
             },
           ),
-          (isLoading && widget.settings.loader != "empty")
+          (isLoading && settings.loader != "empty")
               ? Positioned(
                   top: 0,
                   bottom: 0,
                   right: 0,
                   left: 0,
                   child: Loader(
-                      type: widget.settings.loader,
-                      color: themeProvider.isLightTheme
-                          ? HexColor(widget.settings.loaderColor)
+                      type: settings.loader,
+                      color: themeProvider.isLightTheme!
+                          ? HexColor(settings.loaderColor)
                           : themeProvider.darkTheme.primaryColor))
               : Container()
         ]));
   }
 }
 
-Widget _renderAppBar(context, Settings settings, Page page) {
+AppBar _renderAppBar(context, Settings settings, Page page) {
   var themeProvider = Provider.of<ThemeNotifier>(context);
   return AppBar(
       title: Text(
-        page.title,
+        page.title!,
         style: TextStyle(
             color: Colors.white, fontSize: 22.0, fontWeight: FontWeight.bold),
       ),
@@ -134,10 +134,10 @@ Widget _renderAppBar(context, Settings settings, Page page) {
             begin: Alignment.centerLeft,
             end: Alignment.centerRight,
             colors: <Color>[
-              themeProvider.isLightTheme
+              themeProvider.isLightTheme!
                   ? HexColor(settings.firstColor)
                   : themeProvider.darkTheme.primaryColor,
-              themeProvider.isLightTheme
+              themeProvider.isLightTheme!
                   ? HexColor(settings.secondColor)
                   : themeProvider.darkTheme.primaryColor,
             ],
