@@ -1,50 +1,49 @@
 import 'dart:async';
 import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flyweb/src/data/config.dart';
 import 'package:flyweb/src/helpers/HexColor.dart';
 import 'package:flyweb/src/helpers/SharedPref.dart';
 import 'package:flyweb/src/models/settings.dart';
 import 'package:flyweb/src/pages/HomeScreen.dart';
-import 'package:flyweb/src/pages/OnBoardingScreen.dart';
+import 'package:flyweb/src/pages/onBoarding/OnBoardingScreen.dart';
 import 'package:flyweb/src/repository/settings_service.dart';
-import 'package:flyweb/src/services/theme_manager.dart';
+import 'package:flyweb/src/services/settings_manager.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:onesignal_flutter/onesignal_flutter.dart';
 import 'package:provider/provider.dart';
-import 'package:uni_links/uni_links.dart';
 
 class SplashScreen extends StatefulWidget {
-  final Settings settings;
+  final Settings? localSettings;
 
-  SplashScreen({this.settings});
+  SplashScreen({required this.localSettings});
 
   @override
-  State<StatefulWidget> createState() {
-    // TODO: implement createState
-    return new _SplashScreen(settingsSplach: this.settings);
-  }
+  State<StatefulWidget> createState() => new _SplashScreen();
 }
 
 class _SplashScreen extends State<SplashScreen> {
   final SettingsService settingsService = SettingsService();
   SharedPref sharedPref = SharedPref();
-  String url = "";
-  String onesignalUrl = "";
-  Settings settings = new Settings();
-  Settings settingsSplach = new Settings();
-  bool applicationProblem = false;
-  bool goBoarding = false;
-  StreamSubscription _linkSubscription;
 
-  _SplashScreen({this.settingsSplach});
+  // String url = "";
+  // String onesignalUrl = "";
+  // Settings settings = new Settings();
+
+  bool applicationProblem = false;
+
+  // bool goBoarding = false;
+  // StreamSubscription _linkSubscription;
+
+  _SplashScreen();
 
   @override
   void initState() {
     super.initState();
     initOneSignal();
-    getSettings();
-    loadBoarding();
+    // getSettings();
+    // loadBoarding();
   }
 
   @override
@@ -59,53 +58,58 @@ class _SplashScreen extends State<SplashScreen> {
 
     OneSignal.shared.setRequiresUserPrivacyConsent(true);
 
-    var settings = {OSiOSSettings.autoPrompt: false, OSiOSSettings.promptBeforeOpeningPushUrl: true};
+    var settings = {
+      OSiOSSettings.autoPrompt: false,
+      OSiOSSettings.promptBeforeOpeningPushUrl: true
+    };
 
-    OneSignal.shared.setNotificationReceivedHandler((OSNotification notification) {
+    OneSignal.shared.setNotificationWillShowInForegroundHandler((data) {
       this.setState(() {});
     });
 
-    OneSignal.shared.setNotificationOpenedHandler((OSNotificationOpenedResult result) {
+    OneSignal.shared.setNotificationOpenedHandler(
+        (OSNotificationOpenedResult result) => this.setState(() {}));
+
+    OneSignal.shared
+        .setInAppMessageClickedHandler((OSInAppMessageAction action) {
       this.setState(() {});
     });
 
-    OneSignal.shared.setInAppMessageClickedHandler((OSInAppMessageAction action) {
-      this.setState(() {});
-    });
+    OneSignal.shared
+        .setSubscriptionObserver((OSSubscriptionStateChanges changes) {});
 
-    OneSignal.shared.setSubscriptionObserver((OSSubscriptionStateChanges changes) {});
+    OneSignal.shared
+        .setPermissionObserver((OSPermissionStateChanges changes) {});
 
-    OneSignal.shared.setPermissionObserver((OSPermissionStateChanges changes) {});
+    OneSignal.shared.setEmailSubscriptionObserver(
+        (OSEmailSubscriptionStateChanges changes) {});
 
-    OneSignal.shared.setEmailSubscriptionObserver((OSEmailSubscriptionStateChanges changes) {});
-
+    // todo add [infocusdisplaytype] [iosSetting]
     // NOTE: Replace with your own app ID from https://www.onesignal.com
-    await OneSignal.shared.init('${GlobalConfiguration().getString('appIdOneSignal')}', iOSSettings: settings);
+    await OneSignal.shared.setAppId(
+      '${GlobalConfiguration().getValue('appIdOneSignal')}',
 
-    OneSignal.shared.setInFocusDisplayType(OSNotificationDisplayType.notification);
+      // iOSSettings: settings,
+    );
+
+    // OneSignal.shared
+    //     .setInFocusDisplayType(OSNotificationDisplayType.notification);
 
     bool requiresConsent = await OneSignal.shared.requiresUserPrivacyConsent();
 
     OneSignal.shared.consentGranted(true);
   }
 
-  Future loadBoarding() async {
-    try {
-      var res = await sharedPref.read("boarding");
-      if (res == null) {
-        setState(() {
-          goBoarding = true;
-        });
-      }
-    } on Exception catch (exception) {
-      setState(() {
-        goBoarding = true;
-      });
-    } catch (Excepetion) {
-      setState(() {
-        goBoarding = true;
-      });
-    }
+  Future<bool> loadBoarding() async {
+    var res = await sharedPref.read("boarding").catchError((e) => null);
+    return res == null;
+  }
+
+  ///
+  Future<String?> loadMosqueId() async {
+    String? id = await sharedPref.read("mosqueId").catchError((e) => null);
+
+    return id;
   }
 
   Future<bool> _mockCheckForSession() async {
@@ -113,86 +117,121 @@ class _SplashScreen extends State<SplashScreen> {
     return true;
   }
 
-  getSettings() async {
-    try {
-      Settings _settings = await settingsService.getSettings();
-      sharedPref.save("settings", _settings);
-      this.setState(() {
-        url = _settings.url;
-        if (settings == null) settingsSplach = _settings;
-        settings = _settings;
-        applicationProblem = false;
-      });
-      _mockCheckForSession().then((status) {
-        var future = new Future.delayed(const Duration(milliseconds: 150), _navigateToHome);
-      });
-    } on Exception catch (exception) {
-      this.setState(() {
-        applicationProblem = true;
-      });
-    } catch (Excepetion) {
-      applicationProblem = true;
-    }
-  }
+  // _getSettings() async {
+  //   try {
+  //     Settings _serverSettings = await settingsService.getSettings();
+  //     sharedPref.save("settings", _serverSettings);
+  //     this.setState(() {
+  //       if (settings == null) settingsSplach = _serverSettings;
+  //       settings = _serverSettings;
+  //       applicationProblem = false;
+  //     });
+  //     _mockCheckForSession().then((status) => Future.delayed(
+  //           const Duration(milliseconds: 150),
+  //           _navigateToHome,
+  //         ));
+  //   } on Exception catch (exception) {
+  //     this.setState(() {
+  //       applicationProblem = true;
+  //     });
+  //   } catch (e) {
+  //     applicationProblem = true;
+  //   }
+  // }
 
-  void _navigateToHome() {
-    if (goBoarding && widget.settings.boarding == "1") {
-      sharedPref.save("boarding", "true");
-      Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (BuildContext context) => OnBoardingScreen(onesignalUrl != "" ? onesignalUrl : url, settings)));
+  /// navigates to first screen
+  void _navigateToHome(Settings settings) async {
+    var goBoarding = await loadBoarding();
+    var mosqueId = await loadMosqueId();
+
+    if (mosqueId == null || goBoarding && settings.boarding == "1") {
+      Navigator.of(context).pushReplacement(
+        MaterialPageRoute(
+          builder: (BuildContext context) => OnBoardingScreen(settings),
+        ),
+      );
     } else {
-      Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (BuildContext context) => HomeScreen(onesignalUrl != "" ? onesignalUrl : url, settings)));
+      Navigator.of(context).pushReplacement(
+        MaterialPageRoute(
+          builder: (BuildContext context) => HomeScreen(settings),
+        ),
+      );
     }
   }
 
   Widget build(BuildContext context) {
-    double width = MediaQuery.of(context).size.width;
-    double height = MediaQuery.of(context).size.height;
+    final settingsManager = Provider.of<SettingsManager>(context);
 
-    Color firstColor = (settingsSplach != null && settingsSplach.splash != null && settingsSplach.splash.enable_img == "1")
+    if (settingsManager.settingsLoaded) {
+      // Future.delayed(Duration(milliseconds: 80)).then((value) =>);
+      _navigateToHome(settingsManager.settings);
+    }
+    var settingsSplach = widget.localSettings;
+
+    Color firstColor = (settingsSplach != null &&
+            settingsSplach.splash != null &&
+            settingsSplach.splash!.enable_img == "1")
         ? HexColor("#FFFFFF")
-        : (settingsSplach.splash != null && settingsSplach.splash.firstColor != null && settingsSplach.splash.firstColor != "")
-            ? HexColor(settingsSplach.splash.firstColor)
+        : (settingsSplach!.splash != null &&
+                settingsSplach.splash!.firstColor != null &&
+                settingsSplach.splash!.firstColor != "")
+            ? HexColor(settingsSplach.splash!.firstColor)
             : HexColor('${GlobalConfiguration().getValue('firstColor')}');
 
-    Color secondColor = (settingsSplach != null && settingsSplach.splash != null && settingsSplach.splash.enable_img == "1")
+    Color secondColor = (settingsSplach != null &&
+            settingsSplach.splash != null &&
+            settingsSplach.splash!.enable_img == "1")
         ? HexColor("#FFFFFF")
-        : (settingsSplach.splash != null && settingsSplach.splash.secondColor != null && settingsSplach.splash.secondColor != "")
-            ? HexColor(settingsSplach.splash.secondColor)
+        : (settingsSplach.splash != null &&
+                settingsSplach.splash!.secondColor != null &&
+                settingsSplach.splash!.secondColor != "")
+            ? HexColor(settingsSplach.splash!.secondColor)
             : HexColor('${GlobalConfiguration().getValue('secondColor')}');
 
-    // TODO: implement build
     return Scaffold(
       body: Stack(
         fit: StackFit.expand,
         children: <Widget>[
           Container(
             decoration: BoxDecoration(
-                gradient: LinearGradient(begin: Alignment.topCenter, end: Alignment.bottomCenter, stops: [
-              0,
-              1
-            ], colors: [
-              firstColor,
-              secondColor,
-            ])),
+              gradient: LinearGradient(
+                begin: Alignment.topCenter,
+                end: Alignment.bottomCenter,
+                stops: [0, 1],
+                colors: [
+                  firstColor,
+                  secondColor,
+                ],
+              ),
+            ),
           ),
-          (settingsSplach.splash != null && settingsSplach.splash.enable_img != null && settingsSplach.splash.enable_img == "1")
-              ? Positioned(
-                  top: 0,
-                  right: 0,
-                  child: Image.memory(
-                    Base64Decoder().convert(settingsSplach.splash.img_splash_base64),
-                    fit: BoxFit.cover,
-                    height: height,
-                    width: width,
-                    alignment: Alignment.center,
-                  ),
+          (settingsSplach.splash != null &&
+                  settingsSplach.splash!.enable_img != null &&
+                  settingsSplach.splash!.enable_img == "1")
+              ? /*Image.memory(
+                  Base64Decoder()
+                      .convert(settingsSplach.splash.img_splash_base64),
+                  fit: BoxFit.cover,
+                  height: height,
+                  width: width,
+                  alignment: Alignment.center,
+                )*/
+              Image.asset(
+                  'assets/img/background.png',
+                  fit: BoxFit.cover,
                 )
               : Container(),
-          (settingsSplach.splash != null && settingsSplach.splash.enable_logo != null)
-              ? settingsSplach.splash.enable_logo == "1"
+          (settingsSplach.splash != null &&
+                  settingsSplach.splash!.enable_logo != null)
+              ? settingsSplach.splash!.enable_logo == "1"
                   ? Align(
                       alignment: Alignment.center,
-                      child: Image.memory(Base64Decoder().convert(settingsSplach.splash.logo_splash_base64), height: 150, width: 150),
+                      child: Image.memory(
+                        Base64Decoder().convert(
+                            settingsSplach.splash!.logo_splash_base64!),
+                        height: 150,
+                        width: 150,
+                      ),
                     )
                   : Container()
               : Align(alignment: Alignment.center, child: Config.logo),
@@ -202,21 +241,26 @@ class _SplashScreen extends State<SplashScreen> {
                   right: 0,
                   left: 0,
                   child: Padding(
-                      padding: EdgeInsets.only(top: 50),
-                      child: Column(
-                        children: [
-                          Text(
-                            "System down for maintenance",
-                            textAlign: TextAlign.center,
-                            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20, color: Colors.white),
-                          ),
-                          Text(
-                            "We're sorry, our system is not available",
-                            textAlign: TextAlign.center,
-                            style: TextStyle(fontSize: 18, color: Colors.white),
-                          )
-                        ],
-                      )))
+                    padding: EdgeInsets.only(top: 50),
+                    child: Column(
+                      children: [
+                        Text(
+                          "System down for maintenance",
+                          textAlign: TextAlign.center,
+                          style: TextStyle(
+                              fontWeight: FontWeight.bold,
+                              fontSize: 20,
+                              color: Colors.white),
+                        ),
+                        Text(
+                          "We're sorry, our system is not available",
+                          textAlign: TextAlign.center,
+                          style: TextStyle(fontSize: 18, color: Colors.white),
+                        )
+                      ],
+                    ),
+                  ),
+                )
               : Container()
         ],
       ),
