@@ -1,24 +1,25 @@
 class Splash {
-  String id = "";
-  String firstColor = "";
-  String secondColor = "";
-  String logo_splash = "";
-  String img_splash = "";
-  String enable_logo = "";
-  String enable_img = "";
-  String logo_splash_base64 = "";
-  String img_splash_base64 = "";
+  String? id = "";
+  String? firstColor = "";
+  String? secondColor = "";
+  String? logo_splash = "";
+  String? img_splash = "";
+  String? enable_logo = "";
+  String? enable_img = "";
+  String? logo_splash_base64 = "";
+  String? img_splash_base64 = "";
 
-  Splash(
-      {this.id,
-      this.firstColor,
-      this.secondColor,
-      this.logo_splash,
-      this.img_splash,
-      this.enable_logo,
-      this.enable_img,
-      this.logo_splash_base64,
-      this.img_splash_base64});
+  Splash({
+    this.id,
+    this.firstColor,
+    this.secondColor,
+    this.logo_splash,
+    this.img_splash,
+    this.enable_logo,
+    this.enable_img,
+    this.logo_splash_base64,
+    this.img_splash_base64,
+  });
 
   Splash.fromJson(Map<String, dynamic> json) {
     id = json['id'];
